chore: Remove unused query params [DHIS2-17714]
