--- dhis-2/dhis-api/src/test/java/org/hisp/dhis/trackedentity/TrackedEntityQueryParamsTest.java
@@ -1,60 +1,0 @@
-/*
- * Copyright (c) 2004-2022, University of Oslo
- * All rights reserved.
- *
- * Redistribution and use in source and binary forms, with or without
- * modification, are permitted provided that the following conditions are met:
- * Redistributions of source code must retain the above copyright notice, this
- * list of conditions and the following disclaimer.
- *
- * Redistributions in binary form must reproduce the above copyright notice,
- * this list of conditions and the following disclaimer in the documentation
- * and/or other materials provided with the distribution.
- * Neither the name of the HISP project nor the names of its contributors may
- * be used to endorse or promote products derived from this software without
- * specific prior written permission.
- *
- * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
- * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
- * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
- * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
- * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
- * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
- * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON
- * ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
- * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
- * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
- */
-package org.hisp.dhis.trackedentity;
-
-import static org.hisp.dhis.common.AssignedUserSelectionMode.PROVIDED;
-import static org.junit.jupiter.api.Assertions.assertEquals;
-
-import java.util.Set;
-import org.hisp.dhis.user.User;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.Test;
-
-class TrackedEntityQueryParamsTest {
-
-  private TrackedEntityQueryParams params;
-
-  private User current;
-
-  @BeforeEach
-  void setUp() {
-    current = new User();
-    current.setUid("Kj6vYde4LHh");
-
-    params = new TrackedEntityQueryParams();
-  }
-
-  @Test
-  void testUserWithAssignedUsersGivenCurrentUserAndModeProvidedWithUsers() {
-
-    params.setUserWithAssignedUsers(PROVIDED, Set.of("f1AyMswryyX"));
-
-    assertEquals(PROVIDED, params.getAssignedUserQueryParam().getMode());
-    assertEquals(Set.of("f1AyMswryyX"), params.getAssignedUserQueryParam().getAssignedUsers());
-  }
-}
