+++ dhis-2/dhis-test-integration/src/test/java/org/hisp/dhis/trackedentity/TrackedEntityAttributeStoreIntegrationTest.java
@@ -30,7 +30,6 @@ package org.hisp.dhis.trackedentity;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
 import static org.junit.jupiter.api.Assertions.assertNotNull;
-import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
 import java.util.List;
@@ -39,14 +38,12 @@ import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 import org.hisp.dhis.common.IdentifiableObjectManager;
 import org.hisp.dhis.common.ValueType;
-import org.hisp.dhis.organisationunit.OrganisationUnit;
 import org.hisp.dhis.organisationunit.OrganisationUnitService;
 import org.hisp.dhis.program.Program;
 import org.hisp.dhis.program.ProgramService;
 import org.hisp.dhis.program.ProgramTrackedEntityAttribute;
 import org.hisp.dhis.security.acl.AccessStringHelper;
 import org.hisp.dhis.test.integration.PostgresIntegrationTestBase;
-import org.hisp.dhis.trackedentityattributevalue.TrackedEntityAttributeValue;
 import org.hisp.dhis.tracker.trackedentityattributevalue.TrackedEntityAttributeTableManager;
 import org.hisp.dhis.tracker.trackedentityattributevalue.TrackedEntityAttributeValueService;
 import org.junit.jupiter.api.BeforeEach;
@@ -234,37 +231,5 @@ class TrackedEntityAttributeStoreIntegrationTest extends PostgresIntegrationTest
 
     assertNotNull(attributeIds);
     assertTrue(attributeIds.contains(attributeW.getId()));
-  }
-
-  @Test
-  void shouldValidateUniquenessWhenAttributeIsUnique() {
-    OrganisationUnit organisationUnit = createOrganisationUnit('A');
-    organisationUnitService.addOrganisationUnit(organisationUnit);
-
-    attributeService.addTrackedEntityAttribute(attributeW);
-
-    TrackedEntityType trackedEntityType = createTrackedEntityType('K');
-    TrackedEntityTypeAttribute trackedEntityTypeAttribute = new TrackedEntityTypeAttribute();
-    trackedEntityTypeAttribute.setTrackedEntityAttribute(attributeW);
-    trackedEntityTypeAttribute.setTrackedEntityType(trackedEntityType);
-    trackedEntityType.setTrackedEntityTypeAttributes(List.of(trackedEntityTypeAttribute));
-    trackedEntityTypeService.addTrackedEntityType(trackedEntityType);
-
-    TrackedEntity trackedEntity = createTrackedEntity(organisationUnit);
-
-    trackedEntity.setOrganisationUnit(organisationUnit);
-    trackedEntity.setTrackedEntityType(trackedEntityType);
-
-    manager.save(trackedEntity);
-
-    TrackedEntityAttributeValue trackedEntityAttributeValue = new TrackedEntityAttributeValue();
-    trackedEntityAttributeValue.setAttribute(attributeW);
-    trackedEntityAttributeValue.setTrackedEntity(trackedEntity);
-    trackedEntityAttributeValue.setValue("unique_value");
-    entityAttributeValueService.addTrackedEntityAttributeValue(trackedEntityAttributeValue);
-
-    assertNull(
-        attributeService.validateAttributeUniquenessWithinScope(
-            attributeW, "some_non_unique_value", trackedEntity, organisationUnit));
   }
 }
